refactor: Improve file tree structure
