+++ src/script/components/InputBar/Components/AssetUploadButton/index.ts
