+++ src/script/components/InputBar/Components/ImageUploadButton/ImageUploadButton.test.tsx
@@ -19,7 +19,7 @@
 
 import {render, fireEvent} from '@testing-library/react';
 
-import {ImageUploadButton} from './';
+import {ImageUploadButton} from '.';
 
 jest.mock('../../../Config', () => ({
   Config: {
