+++ src/script/components/InputBar/Components/ImageUploadButton/index.ts
