+++ src/script/components/InputBar/Components/MentionSuggestions/MentionSuggestionsItem.tsx
@@ -26,7 +26,7 @@ import {Icon} from 'Components/Icon';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {handleKeyDown} from 'Util/KeyboardUtil';
 
-import {User} from '../../../entity/User';
+import {User} from '../../../../entity/User';
 
 type MentionSuggestionsItemProps = {
   isSelected: boolean;
