+++ src/script/components/InputBar/Components/PastedFileControls.tsx
@@ -22,7 +22,7 @@ import {FC} from 'react';
 import {Icon} from 'Components/Icon';
 import {t} from 'Util/LocalizerUtil';
 
-import {Config} from '../../Config';
+import {Config} from '../../../Config';
 
 interface PastedFileControlsProps {
   pastedFile: File;
