+++ src/script/components/InputBar/Components/TypingIndicator/TypingIndicator.state.tsx
@@ -19,7 +19,7 @@
 
 import create from 'zustand';
 
-import {User} from '../../../entity/User';
+import {User} from '../../../../entity/User';
 
 type TypingUser = {
   conversationId: string;
