+++ src/script/components/InputBar/Components/TypingIndicator/TypingIndicator.styles.ts
