+++ src/script/components/InputBar/Components/TypingIndicator/TypingIndicator.tsx
