+++ src/script/components/InputBar/Components/TypingIndicator/index.ts
