+++ src/script/components/Modals/PrimaryModal/PrimaryModal.tsx
@@ -23,11 +23,10 @@ import cx from 'classnames';
 
 import {Icon} from 'Components/Icon';
 import {ModalComponent} from 'Components/ModalComponent';
+import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 
 import {usePrimaryModalState, showNextModalInQueue, defaultContent, removeCurrentModal} from './PrimaryModalState';
 import {Action, PrimaryModalType} from './PrimaryModalTypes';
-
-import {initFadingScrollbar} from '../../../ui/fadingScrollbar';
 
 export const PrimaryModalComponent: FC = () => {
   const [inputValue, updateInputValue] = useState<string>('');
