+++ src/script/page/RightSidebar/ConversationParticipants/ConversationParticipants.tsx
@@ -22,6 +22,7 @@ import {FC, useMemo, useState} from 'react';
 import {SearchInput} from 'Components/SearchInput';
 import {UserSearchableList} from 'Components/UserSearchableList';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
+import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 import {t} from 'Util/LocalizerUtil';
 import {sortUsersByPriority} from 'Util/StringUtil';
 
@@ -31,7 +32,6 @@ import {User} from '../../../entity/User';
 import {isServiceEntity} from '../../../guards/Service';
 import {SearchRepository} from '../../../search/SearchRepository';
 import {TeamRepository} from '../../../team/TeamRepository';
-import {initFadingScrollbar} from '../../../ui/fadingScrollbar';
 import {PanelHeader} from '../PanelHeader';
 import {PanelEntity, PanelState} from '../RightSidebar';
