+++ src/script/page/RightSidebar/GuestServicesOptions/GuestServicesOptions.tsx
@@ -21,6 +21,7 @@ import {FC, useState} from 'react';
 
 import {PrimaryModal} from 'Components/Modals/PrimaryModal';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
+import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 import {t} from 'Util/LocalizerUtil';
 
 import {GuestOptions} from './components/GuestOptions';
@@ -31,7 +32,6 @@ import {ConversationRepository} from '../../../conversation/ConversationReposito
 import {Conversation} from '../../../entity/Conversation';
 import {TeamRepository} from '../../../team/TeamRepository';
 import {TeamState} from '../../../team/TeamState';
-import {initFadingScrollbar} from '../../../ui/fadingScrollbar';
 import {PanelHeader} from '../PanelHeader';
 
 interface GuestServicesOptionsProps {
