+++ src/script/util/dom/fadingScrollbar.ts
