+++ src/script/util/dom/resizeObserver.ts
