+++ src/script/util/dom/viewportObserver.ts
